src/scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
anthropic>=0.40.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
lxml>=5.0.0
meilisearch>=0.31.0
